separate qps worker into library with exposed start
