+++ benchmark/worker/main.go
@@ -35,18 +35,8 @@ package main
 
 import (
 	"flag"
-	"fmt"
-	"io"
-	"net"
-	"runtime"
-	"strconv"
-	"time"
 
-	"golang.org/x/net/context"
-	"google.golang.org/grpc"
-	testpb "google.golang.org/grpc/benchmark/grpc_testing"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/grpclog"
+	"google.golang.org/grpc/benchmark/worker/qpsworker"
 )
 
 var (
@@ -54,178 +44,7 @@ var (
 	serverPort = flag.Int("server_port", 0, "port for benchmark server if not specified by server config message")
 )
 
-type byteBufCodec struct {
-}
-
-func (byteBufCodec) Marshal(v interface{}) ([]byte, error) {
-	b, ok := v.(*[]byte)
-	if !ok {
-		return nil, fmt.Errorf("failed to marshal: %v is not type of *[]byte", v)
-	}
-	return *b, nil
-}
-
-func (byteBufCodec) Unmarshal(data []byte, v interface{}) error {
-	b, ok := v.(*[]byte)
-	if !ok {
-		return fmt.Errorf("failed to marshal: %v is not type of *[]byte", v)
-	}
-	*b = data
-	return nil
-}
-
-func (byteBufCodec) String() string {
-	return "bytebuffer"
-}
-
-// workerServer implements WorkerService rpc handlers.
-// It can create benchmarkServer or benchmarkClient on demand.
-type workerServer struct {
-	stop       chan<- bool
-	serverPort int
-}
-
-func (s *workerServer) RunServer(stream testpb.WorkerService_RunServerServer) error {
-	var bs *benchmarkServer
-	defer func() {
-		// Close benchmark server when stream ends.
-		grpclog.Printf("closing benchmark server")
-		if bs != nil {
-			bs.closeFunc()
-		}
-	}()
-	for {
-		in, err := stream.Recv()
-		if err == io.EOF {
-			return nil
-		}
-		if err != nil {
-			return err
-		}
-
-		var out *testpb.ServerStatus
-		switch argtype := in.Argtype.(type) {
-		case *testpb.ServerArgs_Setup:
-			grpclog.Printf("server setup received:")
-			if bs != nil {
-				grpclog.Printf("server setup received when server already exists, closing the existing server")
-				bs.closeFunc()
-			}
-			bs, err = startBenchmarkServer(argtype.Setup, s.serverPort)
-			if err != nil {
-				return err
-			}
-			out = &testpb.ServerStatus{
-				Stats: bs.getStats(false),
-				Port:  int32(bs.port),
-				Cores: int32(bs.cores),
-			}
-
-		case *testpb.ServerArgs_Mark:
-			grpclog.Printf("server mark received:")
-			grpclog.Printf(" - %v", argtype)
-			if bs == nil {
-				return grpc.Errorf(codes.InvalidArgument, "server does not exist when mark received")
-			}
-			out = &testpb.ServerStatus{
-				Stats: bs.getStats(argtype.Mark.Reset_),
-				Port:  int32(bs.port),
-				Cores: int32(bs.cores),
-			}
-		}
-
-		if err := stream.Send(out); err != nil {
-			return err
-		}
-	}
-}
-
-func (s *workerServer) RunClient(stream testpb.WorkerService_RunClientServer) error {
-	var bc *benchmarkClient
-	defer func() {
-		// Shut down benchmark client when stream ends.
-		grpclog.Printf("shuting down benchmark client")
-		if bc != nil {
-			bc.shutdown()
-		}
-	}()
-	for {
-		in, err := stream.Recv()
-		if err == io.EOF {
-			return nil
-		}
-		if err != nil {
-			return err
-		}
-
-		var out *testpb.ClientStatus
-		switch t := in.Argtype.(type) {
-		case *testpb.ClientArgs_Setup:
-			grpclog.Printf("client setup received:")
-			if bc != nil {
-				grpclog.Printf("client setup received when client already exists, shuting down the existing client")
-				bc.shutdown()
-			}
-			bc, err = startBenchmarkClient(t.Setup)
-			if err != nil {
-				return err
-			}
-			out = &testpb.ClientStatus{
-				Stats: bc.getStats(false),
-			}
-
-		case *testpb.ClientArgs_Mark:
-			grpclog.Printf("client mark received:")
-			grpclog.Printf(" - %v", t)
-			if bc == nil {
-				return grpc.Errorf(codes.InvalidArgument, "client does not exist when mark received")
-			}
-			out = &testpb.ClientStatus{
-				Stats: bc.getStats(t.Mark.Reset_),
-			}
-		}
-
-		if err := stream.Send(out); err != nil {
-			return err
-		}
-	}
-}
-
-func (s *workerServer) CoreCount(ctx context.Context, in *testpb.CoreRequest) (*testpb.CoreResponse, error) {
-	grpclog.Printf("core count: %v", runtime.NumCPU())
-	return &testpb.CoreResponse{Cores: int32(runtime.NumCPU())}, nil
-}
-
-func (s *workerServer) QuitWorker(ctx context.Context, in *testpb.Void) (*testpb.Void, error) {
-	grpclog.Printf("quiting worker")
-	s.stop <- true
-	return &testpb.Void{}, nil
-}
-
 func main() {
-	grpc.EnableTracing = false
-
 	flag.Parse()
-	lis, err := net.Listen("tcp", ":"+strconv.Itoa(*driverPort))
-	if err != nil {
-		grpclog.Fatalf("failed to listen: %v", err)
-	}
-	grpclog.Printf("worker listening at port %v", *driverPort)
-
-	s := grpc.NewServer()
-	stop := make(chan bool)
-	testpb.RegisterWorkerServiceServer(s, &workerServer{
-		stop:       stop,
-		serverPort: *serverPort,
-	})
-
-	go func() {
-		<-stop
-		// Wait for 1 second before stopping the server to make sure the return value of QuitWorker is sent to client.
-		// TODO revise this once server graceful stop is supported in gRPC.
-		time.Sleep(time.Second)
-		s.Stop()
-	}()
-
-	s.Serve(lis)
+	qpsworker.RunBenchmarkWorker(*driverPort, *serverPort, nil)
 }
